Make RuleSetLocator private
